requirements.txt
pillow
pypdf